feat(microcache): Create new system of cache
POC with two new caches, edit getter/setter in gsh and api
usefull accordingly

+++ Holovibes/sources/gui/windows/AdvancedSettingsWindow.cc
@@ -45,7 +45,7 @@ void AdvancedSettingsWindow::closeEvent(QCloseEvent* event) { emit closed(); }
 void AdvancedSettingsWindow::set_ui_values()
 {
     api::set_file_buffer_size(static_cast<int>(ui.FileBSSpinBox->value()));
-    api::set_input_buffer_size(static_cast<int>(ui.InputBSSpinBox->value()));
+    api::set_input_buffer_size(ui.InputBSSpinBox->value());
     api::set_record_buffer_size(static_cast<int>(ui.RecordBSSpinBox->value()));
     api::set_output_buffer_size(static_cast<int>(ui.OutputBSSpinBox->value()));
     api::set_time_transformation_cuts_output_buffer_size(static_cast<int>(ui.Cuts3DBSSpinBox->value()));
